Canva pour comprendre les polygons donner dans le ground truth

+++ app/src/main/java/com/example/mlkitevaluation/vo/TotalTextGroundTruth.kt
@@ -1,8 +1,8 @@
 package com.example.mlkitevaluation.vo
 
 data class TotalTextGroundTruth(
-    val x: List<List<Int>>,
-    val y: List<List<Int>>,
+    val x: List<List<Float>>,
+    val y: List<List<Float>>,
     val orientation: TotalTextOrientation,
     val transcription: String
 )
